refactor dashboard table
