+++ css/src/general-page.css
@@ -248,33 +248,6 @@ body.toplevel_page_wpseo_dashboard {
 		}
 	}
 
-
-	.yst-site-kit-widget-table td:first-child {
-		counter-increment: row-number;
-		@apply yst-relative;
-	}
-
-	.yst-site-kit-widget-table td:first-child div {
-		@apply yst-text-slate-900 yst-font-medium;
-	}
-	.yst-site-kit-widget-table td:first-child div, .yst-site-kit-widget-table th:first-child div{
-		@apply yst-ps-3;
-	}
-
-	.yst-site-kit-widget-table td:first-child::before {
-		content: counter(row-number) ". ";
-		@apply yst-absolute yst-left-0 yst-text-slate-500 yst-font-normal;
-	}
-
-	/* Initialize the counter at the parent element level */
-	.yst-site-kit-widget-table tbody {
-		counter-reset: row-number;
-	}
-
-	.yst-site-kit-widget-table th:last-child div {
-		@apply yst-text-end;
-	}
-
 	/* RTL */
 
 	&.rtl {
